# [purs ide] Parses modules in parallel