Fixes #2783 

Empty project with just Halogen:
```
before: 
[perf] Command Load took 3092.80ms
after:  
[perf] Command Load took 1994.84ms
```


Slamdata:
```
before:
  λ purs ide server --log-level perf "src/**/*.purs" "bower_components/*/src/**/*.purs"
  [perf] Command Load took 12572.22ms
after:
  λ purs ide server --log-level perf "src/**/*.purs" "bower_components/*/src/**/*.purs"
  [perf] Command Load took 7251.82ms
```